+++ CDLF/include/ConvolutionLayer.h
@@ -15,7 +15,7 @@
  *       * indicate convolution
  *
  * Notes:
- * 1  in convolution layer, we do not consider bias, as there is a separate BiasLayer for use;
+ * 1  in convolution layer, we do not consider bias, as there is a separate LinearLayer for use;
  * 2  Size changes: |Y| = (|X|-|W|)/stride + 1, in their different dimension;
  * 3  it is a good design if all numFilter is odd;
  * 4  Currently we supports 1D, 2D, 3D, 4D, 5D, 6D convolution; It is easy to extend to 7D or more.
@@ -25,7 +25,7 @@
 
 class ConvolutionLayer :  public ConvolutionBasicLayer {
 public:
-    ConvolutionLayer(const int id, const string& name, Layer* prevLayer, const vector<long>& filterSize,
+    ConvolutionLayer(const int id, const string& name, Layer* prevLayer, const vector<int>& filterSize,
                      const int numFilters=1, const int stride=1);
     ~ConvolutionLayer();
