+++ CDLF/include/LossLayer.h
@@ -33,7 +33,7 @@ public:
 
     void setGroundTruth( const Tensor<float>& groundTruth);
     void setGroundTruth( const Tensor<unsigned  char>& groundTruth);
-    virtual  long getNumParameters();
+    virtual  int getNumParameters();
 
     virtual  void save(const string& netDir);
     virtual  void load(const string& netDir);
