+++ CDLF/include/SoftmaxLayer.h
@@ -26,7 +26,7 @@ public:
     virtual  void forward();
     virtual  void backward(bool computeW, bool computeX = true);
     virtual  void updateParameters(const float lr, const string& method, const int batchSize=1);
-    virtual  long getNumParameters();
+    virtual  int getNumParameters();
     virtual  void save(const string& netDir);
     virtual  void load(const string& netDir);
     virtual  void saveStructLine(FILE* pFile);
